legal: Added FOSS Licenses legal section to AdwAboutDialog
If you're wondering why I didn't do this the 'ideal' way, or why I did
this in the first place: it is just to show respect to the developers
who made all the underlying libraries.

I wanted to include every single name of every major contributor to each
library that Album uses. :)