library_list_model: Hide GFileInfo list items that represent directories
Only thing I did not fix here (just yet) is that there is a segfault on
exiting the program. but it looks like the view is handling this
'hidden' items illusion well.

We are hiding GFileInfo list items that represent directories since we
only want to display photo/video files from your library and not your
folders. :P