l10n/build-aux: Fix generate-potfile.sh && misc l10n edit
The other misc edit is making the error page in the library-view.ui have
a translatable placeholder description for its status page so that if,
by any case, this appears to a user, they know to report the issue.

I know this is an odd commit, but I'm trying to avoid cluttering the git
commit history full of nit picking commits so I did a 2 in 1 here.