thumbnails: Set ffmpeg loglevel to err to fix fake IO errors
Sorry the above is worded badly. Basically for some reason ffmpeg always
wants to print out to stderr so even on successful thumbnails generated,
our code thinks an error occurred and raises an IO error.

Setting the loglevel to error on ffmpeg allows for it to print
absolutely nothing unless an error actually ocurred in ffmpeg.