thumbnails/utils: Cache thumbnails by their metadata's MD5 fingerprint
We're now reading the input file's metadata and storing it into a data
structure that can be serialized by serde into a JSON byte array, that
is then sent to a MD5 hasher to generate an MD5 fingerprint hash for
that file's metadata information.

This will allow the program to regenerate the file thumbnail if it is
edited by the user, which is vital for future features of Albums.