library_view: Woops, accidentally had broken mobile gallery grid zoom
When I was writing the block {} for the hardware_accel property, I
initially put it in the line for the grid_desktop_zoom property, then I
moved it to the right one and reset the grid_desktop_zoom property to
true, but its original value was actually false. Human error. :P