preferences/gsettings: Disable hardware acceleration by default
Yeahh this is definitely an experimental feature for now. On desktop, it
actually works well although it does fail some images, but on mobile it
seems to be having issues, though it could be an issue on my phone's
system setup or the Danctnix ALARM distro. Anyways, too risky to ship.