gschema/app: Set 'fresh-cache' to true if cache missing
Moved cache check logic from main.rs to application/mod.rs so that we
can have access to the `GSettings` instance for accessing our gschema
settings.

Also moved the logic that checks and unsets the 'fresh-cache' flag from
the library/mod.rs:load_library() to the list model's `models_loaded`
callback so that the user is shown the `AdwToast` **after** the
`LibraryListModel` is done enumerating the user's library.