chore/refactor: Untangle photo grid view from library view
Previously, the photo grid view in the library view was basically a
singleton, if that is the right term. That grid view was the only
instance of a photo grid view in Albums, and if we needed another one
(for viewing albums in the future) we would have to rewrite code.

This refactor moves all the photo grid view logic / widget building to
`AlbumsMediaGridView`, and also moves `AlbumsMediaCell` into a submodule
of `media_grid` which holds the new subclass.

This refactor is in preparation of the albums feature.