library/viewer: Toggle fullscreen and moved details sidebar right
The details sidebar on the media viewer has been moved from the left, to
the right, in order to avoid overriding the swipe gesture on the
lefthand side of the screen on mobile for popping the navigation page
from the navigation view.

Also added a toggle fullscreen button for those who want to maximize
screen space on their mobile device when viewing pictures and videos.
This helps a lot when you are viewing media on landscape mode.