app: Renamed to codename "Memories", see description
The GNOME incubator submission info / app criteria mentions that any
applications considering to join the program should be named under a
codename, and are later branded a generic name if accepted into GNOME
core. This means we should not occupy the name "Albums". (Even though if
Memories were to make it to Incubator, it would be most likely planned
to be named "Photos")